ARC-3493: Clean up the code a bit more
