+++ src/modules/account/components/MaterialRequestDetailBlade/MaterialRequestContentInfo.module.scss
@@ -1,0 +1,51 @@
+@use "../../../../styles/abstracts" as *;
+
+.p-material-request-detail {
+	&__content-block {
+		padding: $spacer-sm 0;
+		display: flex;
+		flex-direction: column;
+		border-bottom: 1px solid $silver;
+		margin-block: 0;
+
+		&:first-child {
+			padding-top: 0;
+		}
+
+		&:last-child {
+			border-bottom: none;
+			padding-bottom: 0;
+		}
+
+		dt, dd {
+			margin: 0;
+		}
+
+		&-title {
+			font-size: $font-size-md;
+			line-height: $font-size-base;
+			font-weight: $font-weight-bold;
+			padding-bottom: $spacer-xs;
+		}
+
+		&-subtitle {
+			font-size: $font-size-base;
+			line-height: $font-size-lg;
+			padding-bottom: $spacer-xxs;
+		}
+
+		&-value {
+			color: $grey;
+			padding: 0;
+			font-size: $font-size-sm;
+			line-height: $font-size-base;
+		}
+
+		&-media {
+			width: 100%;
+			aspect-ratio: 16/9;
+			object-fit: cover;
+			object-position: center;
+		}
+	}
+}
