+++ src/modules/account/components/MaterialRequestDetailBlade/material-request-detail-blade.consts.tsx
@@ -1,8 +1,24 @@
+import {
+	type MaterialRequest,
+	type MaterialRequestEventType,
+	MaterialRequestStatus,
+} from '@material-requests/types';
 import type { TabProps } from '@meemoo/react-components';
 import { Icon } from '@shared/components/Icon';
 import { IconNamesLight } from '@shared/components/Icon/Icon.enums';
 import { tText } from '@shared/helpers/translate';
 import { MaterialRequestDetailBladeTabs } from './material-request-detail-blade.types';
+
+export const isLatestEventStatus = (
+	materialRequest: MaterialRequest,
+	status: MaterialRequestEventType
+): boolean => {
+	return (
+		materialRequest.status === MaterialRequestStatus.PENDING &&
+		materialRequest.history.length > 0 &&
+		materialRequest.history.at(materialRequest.history.length - 1)?.messageType === status
+	);
+};
 
 /**
  * Tabs
